Update README.md
Overview

The plots2 Readme has become too long since its introduction. We need to modularize the readme and transfer the contents in separate github wiki pages.

Objective

A long readme is really hard to read and also it is hard to search and traverse any topic. So it is better to create separate wiki pages in the github repository. Separate links can be given in the readme with some introduction.

Requirements

Learn git for getting started
Install the repository in your local
Create github wiki pages for each section
Update readme with the links to wiki pages and submit a pull request

Expected Outcome

Github wiki pages for various sections in readme
Suggest changes in readme
A pull request to update the readme

https://github.com/publiclab/plots2#publiclaborg

GCI 2016 Emily Ong