Explicitly set ruby version (#2570)
* Explicitly set Ruby version
This makes it easier for development, this way RVM autmatically picks the Ruby version to use. If the dev hasn't got this version installed, RVM helps him install it and in cases where we have different versions of ruby, we don't have to set the version number every time once we enter Plot2. Time saver

* Update readme

* Update Ruby version