Fixes number of "wiki edits" displayed on Dashboard
The query used to count recent wiki edits was not limited to counting
only wiki pages, published ones, and distint pages (i.e. do not count
multiple revisions of the same page).

This resulted in a major difference between what the person sees on
their dashboard vs. when they scroll through the list on `/wiki`, which
shows a more accurate number of recent changes.

This commit fixes the problem and adds a test that verifies it.