Move and reformat 'by likes' and 'by views' buttons
The 'by likes' and 'by views' buttons were moved to the right side next to the 'by type' button, and their widths were increased (by removing class btn-sm) to match the 'by type' button.
The text on the 'by type' was changed from 'By Type' to 'by type,' in order to be consistent in capitalization with the 'by likes' and 'by views' buttons.
'# posts by # contributors' on the left side was corrected to use the correct method for the contributor count.