Fixed notes_controller so that notes are sorted by likes (#6919)
* Fixed notes_controller so that notes are sort by likes

The notes were being ordered through their 'nid' which I assume means node id instead of their 'cached_likes' count. I think this is the solution to problem and it worked on my local machine. I looked at schema for the nodes and think this is what they need to be ordered by.

* now /notes/recent will be sorted from most recent to least