REFACTOR: post_test.rb [Travis optimization]
The most important improvement is the removal of assert_page_reloads statement.
This block of code never executed and just delayed Travis for ~1min every build.

The second improvement is the addition of the "setup" function.
Instead of writing login logic in every test, now before every test the
"setup" function will run and log in the user.

Part of #7272