FIX: Wiki revisions test (#7396)
This test has already been fixed before but recently it has failed again
in #7319. The problem might be related to our selection of "Revert"
button.

Inside of the button we have an empty space before the text
and a font-awesome icon(<i>).
We were selecting the button with #row0 and the text of "Revert" which
probably caused the issue.

I've added the btn-revert class to avoid this strange behaviour.