Update lines 10 and 13
I left lines 14, 17, 24, 26, 32, 37, 40, 43 and 46 unchanged as they did not seem to meet the criteria for the change request.  I was not sure if the font class was to be added if it did not previously exist on that line. Just let me know if I need to add them.

Thanks,
SeriousHyena